feat(web): scaffold vue-router with /generate route
