+++ web/package.json
@@ -12,7 +12,8 @@
     "fmt:check": "prettier --check ."
   },
   "dependencies": {
-    "vue": "^3.5.13"
+    "vue": "^3.5.13",
+    "vue-router": "^4.4.0"
   },
   "devDependencies": {
     "@tailwindcss/vite": "^4.2.0",
