+++ web/src/pages/GeneratePage.vue
@@ -1,0 +1,10 @@
+<script setup lang="ts">
+// Scaffold for Task 1. Composer / RunningStrip / feed land in later tasks.
+</script>
+
+<template>
+  <div class="mx-auto max-w-[1800px] px-4 py-6 sm:px-6 lg:px-10">
+    <h1 class="text-2xl font-semibold text-slate-100">Generate</h1>
+    <p class="mt-2 text-sm text-slate-400">Coming online in follow-up tasks.</p>
+  </div>
+</template>
